Arreglado Venta
Arregle error que corrompia clase cliente ya se carga venta y interactua con el resto de clases faltan validaciones y ver si sumosmo funcionalidad

+++ Cliente.cpp
@@ -10,12 +10,12 @@ using namespace std;
 
 
 
-void Cliente::setTotalGastado(int totalGastado)
+void Cliente::setTotalGastado(float totalGastado)
 {
     _totalGastado=totalGastado;
 }
 
-int Cliente::getTotalGastado() const
+float Cliente::getTotalGastado() const
 {
     return _totalGastado;
 }
